Incorrect key for probing the originToReference maps. Must use a long to avoid testing equality word/object.

+++ com.oracle.max.tele.vm/src/com/sun/max/tele/object/FixedObjectRemoteReferenceManager.java
@@ -75,7 +75,7 @@ public final class FixedObjectRemoteReferenceManager extends AbstractVmHolder im
 
     public ObjectStatus objectStatusAt(Address origin) throws TeleError {
         TeleError.check(objectRegion.memoryRegion().contains(origin), "Location is outside region");
-        final WeakReference<RemoteReference> weakRef = originToReference.get(origin);
+        final WeakReference<RemoteReference> weakRef = originToReference.get(origin.toLong());
         if (weakRef != null) {
             final RemoteReference knownReference = weakRef.get();
             if (knownReference != null) {
